server/requirements.txt
aniso8601==8.1.1
async-generator==1.10
attrs==20.3.0
certifi==2020.12.5
click==7.1.2
colorama==0.4.3
dnspython==2.1.0
fire==0.3.1
Flask==1.1.2
flask-restx==0.2.0
gitdb==4.0.5
GitPython==3.1.7
itsdangerous==1.1.0
jill==0.6.14
Jinja2==2.11.3
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-server==0.1.1
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
loguru==0.5.3
MarkupSafe==1.1.1
nbdime==2.0.0
psycopg2==2.8.6
Pygments==2.6.1
PyJWT==2.0.1
pymongo==3.11.3
pyrsistent==0.17.3
python-dotenv==0.15.0
python-environ==0.4.54
python-gnupg==0.4.6
pytz==2021.1
scikit-learn==0.23.2
scipy==1.5.2
semantic-version==2.8.5
six==1.15.0
sklearn==0.0
smmap==3.0.4
termcolor==1.1.0
threadpoolctl==2.1.0
voila==0.1.21
Werkzeug==1.0.1
wget==3.2
